PasswordGenerationLogic moved to separated file
